+++ docs/CHANGELOG.md
@@ -19,6 +19,8 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 * `SystemDesc` proc macro derive to simplify defining `SystemDesc`s. ([#1780])
 * `UiButtonData` is now exported from `amethyst_ui` and can be used for custom widgets. ([#1859])
 * Add an audio subchapter to the pong chapter. ([#1842])
+* Add `DispatcherOperation` to store dispatcher build logic, which can be executed lazily. ([#1870])
+
 
 ### Changed
 
@@ -29,6 +31,7 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 [#1780]: https://github.com/amethyst/amethyst/pull/1780
 [#1859]: https://github.com/amethyst/amethyst/pull/1859
 [#1842]: https://github.com/amethyst/amethyst/pull/1842
+[#1870]: https://github.com/amethyst/amethyst/pull/1870
 
 ## [0.12.0] - 2019-07-30